+++ src/inlineSuggestions/registerHandlers.ts
@@ -2,14 +2,12 @@ import {
   commands,
   Disposable,
   languages,
-  Range,
   TextEditor,
   TextEditorSelectionChangeEvent,
   TextEditorSelectionChangeKind,
   window,
   workspace,
 } from "vscode";
-import fetch from "node-fetch";
 import { Capability, isCapabilityEnabled } from "../capabilities/capabilities";
 import {
   ACCEPT_INLINE_COMMAND,
@@ -35,8 +33,7 @@ import {
   isInlineSuggestionProposedApiSupported,
   isInlineSuggestionReleasedApiSupported,
 } from "../globals/versions";
-
-// export const decorationType = window.createTextEditorDecorationType({});
+import highlightStackAttributions from "./highlightStackAttributions";
 
 function isSnippetAutoTriggerEnabled() {
   return isCapabilityEnabled(Capability.SNIPPET_AUTO_TRIGGER);
@@ -171,46 +168,9 @@ function registerEscapeHandler(): Disposable {
   });
 }
 
-function registerAttributionHandler(): Disposable {
-  console.log("REGISTERING ATTRIBUTION HANDLER");
-  
-  return commands.registerTextEditorCommand(`${ATTRIBUTION_COMMAND}`, ({document, selection}) => {
-    const text = document.getText();
-    const url = "https://stack-dev.dataportraits.org/overlap";
-    const body = { document: text };
-    
-    void fetch(url, {
-      method: "POST",
-      body: JSON.stringify(body),
-      headers: { "Content-Type": "application/json" },
-    }).then(async resp => {
-      const json = await resp.json() as any as {spans: [[number, number]]}      
-      const {spans} = json
-
-      // find the biggest span
-      const [startChar, endChar] = spans.reduce((acc, curr) => {
-        const [accStart, accEnd] = acc;
-        const [currStart, currEnd] = curr;
-        const accLength = accEnd - accStart;
-        const currLength = currEnd - currStart;
-        if (currLength > accLength) {
-          return curr;
-        }
-        return acc;
-      }, [0, 0]);
-
-      const decorations = [{range: new Range(document.positionAt(startChar), document.positionAt(endChar)), hoverMessage: "This code is in the stack!"}]
-
-      const decorationType = window.createTextEditorDecorationType({
-        color: 'red',
-        textDecoration: 'underline',
-      });
-      window.activeTextEditor?.setDecorations(decorationType, decorations);
-
-      setTimeout(() => {
-        window.activeTextEditor?.setDecorations(decorationType, []);
-      }, 10000);
-    })
+function registerAttributionHandler(): Disposable {  
+  return commands.registerTextEditorCommand(`${ATTRIBUTION_COMMAND}`, () => {
+    void highlightStackAttributions();
   });
 }
